add some unittests
