+++ feeluown/app/__init__.py
@@ -1,8 +1,13 @@
 from .app import App, AppMode, create_app
+from .args import init_args_parser
+from .config import create_config
 
 
 __all__ = (
     'App',
     'AppMode',
     'create_app',
+
+    'init_args_parser',
+    'create_config',
 )
