+++ feeluown/cli/__init__.py
@@ -1,2 +1,2 @@
-from .cli import setup_cli_argparse, climain, oncemain  # noqa
+from .cli import init_args_parser, climain, oncemain  # noqa
 from .cli import Client, Request  # noqa
